requirements.txt
requests
beautifulsoup4
pandas
sqlalchemy
python-dotenv
psycopg2-binary
geopy
mysqlclient
datetime
ThreadPoolExecutor 